fix: jsonMenuNested modals not working
The channel features introduce a new formTemplate.
JsonMenuNested modals are defined in the form_end, but this was not called anymore!

Because @EMSCore/form/forms.html.twig is more important it should come at the end of the array!

Symfony documentation:
You can pass multiple themes to this option because sometimes form themes only redefine a few elements. This way, if some theme doesn’t override some element, Symfony looks up in the other themes.

The order of the themes in the twig.form_themes option is important. Each theme overrides all the previous themes, so you must put the most important themes at the end of the list.