fix: version tag from rawData
Setting meta fields on a revision should first look at the rawData for setting the version tag (Major/Minor,...). Because when we migrate only the rawData is set on the revision. Otherwise all revision will become the default versionTag.

On replaceData we should also call setMetaFields for updating incorrect versionTags from the api.