fix: UploadedFileController not defined in controllers.xml (#1059)
In the past it was loaded because of the prototype, which was not okay:
https://github.com/ems-project/EMSCoreBundle/commit/2e0afddf37d13df8b9f307854a5b8f5af84dd35b

Error:
The controller for URI "/publisher/uploaded-file/" is not callable: Controller "EMS\CoreBundle\Controller\UploadedFileController" has required constructor arguments and does not exist in the container. Did you forget to define the controller as a service?