fix: app.twig_extension needs to be public
Because used in the dataService with $this->container->get('app.twig_extension');
Needs to be refactored in the feature

The "app.twig_extension" service or alias has been removed or inlined when the container was compiled. You should either make it public, or stop using the container directly and use dependency injection instead.